Improve UX of OCR-recognized images
Rather than dumping the JSON, let's output a USWDS style infobox with
the contents rendered a bit nicer for non-techies.

This also institutes an arbitrary "45" threshold of confidence to
consider the OCR result "legible". Although people on the internet say
that confidence really should be 95+, this is a lower bound that will
hopefully not raise too many false positives.